aphasia-api/requirements.txt
fastapi
uvicorn
torch
transformers
librosa
numpy
scipy
fastdtw
openai-whisper
loguru
python-multipart
pydantic>=2,<3
